Add support for picking up and removing items
Note that this involved a substantial refactor to item behaviors that I expect will propagate across a couple other parts of our domain as well. Now the behavior is a type that contains the relevant functions. Each behavior will have a corresponding ID which keys it into an item behavior lookup table, and the ItemData contains the corresponding behavior ID which is used to call the correct functions given the item ID.